moving fee calculator method to permit fee helper
